"Add main program files and configure project setup"
This commit includes the addition of two new java classes, 'Main.java' and 'GitCommitHash.java'. 'Main.java' is a placeholder class for the main program, while 'GitCommitHash.java' includes a method to fetch Git commit hash of the latest commit. This is useful for having a reference to the version of the code at runtime.
A '.gitignore' file has been added that includes patterns for Java and Maven related files that should not be tracked.
A 'pom.xml' file is provided to configure Maven project settings and dependencies. Here, a dependency on the 'jgit' library is included which is used in GitCommitHash.java.